Detect entity interaction & debug rendering
- Press XOR + MEM0 to activate debug display

+++ Software/src/games/mage/mage_map.cpp
@@ -194,7 +194,7 @@ uint8_t MageMap::LayerCount() const
 	return layerCount;
 }
 
-uint16_t MageMap::EntityCount() const
+uint8_t MageMap::EntityCount() const
 {
 	return entityCount;
 }
